ip match successful

+++ src/cn/edu/sjtu/front/iputils/IpMatcher.java
@@ -9,7 +9,7 @@ package cn.edu.sjtu.front.iputils;
  */
 interface IpMatcher {
 
-	public abstract Ipv4NetAddr ipMatch(int ipv4Addr);
+	public abstract Ipv4NetAddr ipMatch(long ipv4Addr);
 	public abstract Ipv4NetAddr ipMatch(Ipv4NetAddr v4NetAddr);
 	
 }
